+++ python/benchmarks/real_batch_functions.py
@@ -9,12 +9,13 @@ As of the Change Date specified in that file, in accordance with the Business So
 import os
 import numpy as np
 import pandas as pd
-from arcticdb.util.environment_setup_v1 import SetupMultipleLibraries, Storage
+from arcticdb.util.environment_setup import LibraryManager, LibraryPopulationPolicy, LibraryType, Storage, populate_library
 from arcticdb.util.utils import TimestampNumber
 from arcticdb.version_store.library import Library, ReadRequest, WritePayload
+from benchmarks.common import AsvBase
 
 
-class AWSBatchBasicFunctions:
+class AWSBatchBasicFunctions(AsvBase):
     """
     This is similar test to :class:`BatchBasicFunctions`
     Note that because batch functions are silent we do check if they work correctly along with 
@@ -29,104 +30,124 @@ class AWSBatchBasicFunctions:
 
     timeout = 1200
 
-    SETUP_CLASS = (SetupMultipleLibraries(storage=Storage.AMAZON, 
-                                                      prefix="BASIC_BATCH")
-                   .set_params([[500, 1000], [25_000, 50_000]])
-                   .set_number_columns(10)
-                   )
-    params = SETUP_CLASS.get_parameter_list()
+    params = [[5, 10], [250, 500]]
     param_names = ["num_symbols", "num_rows"]
 
+    library_manager = LibraryManager(Storage.LMDB, "BASIC_BATCH")
+
+    number_columns = 10
+    initial_timestamp = pd.Timestamp("10-11-1978")
+    freq = 's'
+
+    def get_library_manager(self) -> LibraryManager:
+        return AWSBatchBasicFunctions.library_manager
+    
+    def get_population_policy(self) -> LibraryPopulationPolicy:
+        lpp = LibraryPopulationPolicy(AWSBatchBasicFunctions.params, self.get_logger())
+        lpp.set_columns(AWSBatchBasicFunctions.number_columns)
+        lpp.use_auto_increment_index()
+        return lpp
+    
     def setup_cache(self):
-        set_env = AWSBatchBasicFunctions.SETUP_CLASS
-        set_env.setup_environment()
-        info = set_env.get_storage_info()
-        # NOTE: use only logger defined by setup class
-        set_env.logger().info(f"storage info object: {info}")
-        return info
+        manager = self.get_library_manager()
+        policy = self.get_population_policy()
+        number_symbols_list, number_rows_list = AWSBatchBasicFunctions.params
+        for number_symbols in number_symbols_list:
+            lib_suffix = number_symbols
+            if not manager.has_library(LibraryType.PERSISTENT, lib_suffix):
+                for number_rows in number_rows_list:
+                    policy.set_parameters([number_rows] * lib_suffix)
+                    # the name of symbols during generation will have now 2 parameters:
+                    # the index of symbol + number of rows
+                    # that allows generating more than one symbol in a library
+                    policy.set_symbol_fixed_str(number_rows) 
+                    populate_library(manager, policy, LibraryType.PERSISTENT, lib_suffix)
+        manager.log_info() # Always log the ArcticURIs 
 
-    def teardown(self, storage_info, num_symbols, num_rows):
-        ## Delete own modifyable library
-        self.setup_env.delete_modifiable_library(os.getpid())
+    def teardown(self, num_symbols, num_rows):
+        # We could clear the modifiable libraries we used
+        self.get_library_manager().clear_all_modifiable_libs_from_this_process()
 
-    def setup(self, storage_info, num_symbols, num_rows):
-        self.setup_env = SetupMultipleLibraries.from_storage_info(storage_info)
-        
-        self.lib: Library = self.setup_env.get_library(num_symbols)
-        self.setup_env.logger().info(f"Library {self.lib}") 
-        self.setup_env.logger().debug(f"Symbols {self.lib.list_symbols()}") 
+    def setup(self, num_symbols, num_rows):
+        self.manager = self.get_library_manager()
+        self.population_policy = self.get_population_policy()
+        # We use the same generator as the policy
+
+        self.lib: Library = self.manager.get_library(LibraryType.PERSISTENT, num_symbols)
+        self.write_lib: Library = self.manager.get_library(LibraryType.MODIFIABLE, num_symbols)
+        self.get_logger().info(f"Library {self.lib}") 
+        self.get_logger().debug(f"Symbols {self.lib.list_symbols()}") 
         
         # Get generated symbol names
         self.symbols = []
         for num_symb_idx in range(num_symbols):
-            sym_name = self.setup_env.get_symbol_name(num_symb_idx, num_rows, self.setup_env.default_number_cols)
+            # the name is constructed of 2 parts index + number of rows
+            sym_name = self.population_policy.get_symbol_name(num_symb_idx, num_rows)
             if not self.lib.has_symbol(sym_name):
-                self.setup_env.logger().error(f"symbol not found {sym_name}") 
+                self.get_logger().error(f"symbol not found {sym_name}") 
             self.symbols.append(sym_name)
 
         #Construct read requests (will equal to number of symbols)
         self.read_reqs = [ReadRequest(symbol) for symbol in self.symbols]
 
-        #Construct read requests based on 2 colmns, not whole DF (will equal to number of symbols)
-        COLS = self.setup_env.generate_dataframe(0, self.setup_env.default_number_cols).columns[2:4]
-        self.read_reqs_with_cols = [ReadRequest(symbol, columns=COLS) for symbol in self.symbols]
-
         #Construct dataframe that will be used for write requests, not whole DF (will equal to number of symbols)
-        self.df: pd.DataFrame = self.setup_env.generate_dataframe(num_rows, self.setup_env.default_number_cols)
+        self.df = self.population_policy.df_generator.get_dataframe(num_rows, AWSBatchBasicFunctions.number_columns)
+
+        #Construct read requests based on 2 colmns, not whole DF (will equal to number of symbols)
+        COLS = self.df.columns[2:4]
+        self.read_reqs_with_cols = [ReadRequest(symbol, columns=COLS) for symbol in self.symbols]
 
         #Construct read request with date_range
         self.date_range = self.get_last_x_percent_date_range(num_rows, 0.05)
         self.read_reqs_date_range = [ReadRequest(symbol, date_range=self.date_range) for symbol in self.symbols]
-        
-        ## Make sure each process has its own write library
-        self.fresh_lib: Library = self.setup_env.get_modifiable_library(os.getpid())
 
     def get_last_x_percent_date_range(self, num_rows, percents):
         """
         Returns a date range tuple selecting last X% of rows of dataframe
         pass percents as 0.0-1.0
         """
-        freq = self.setup_env.index_freq
-        start = TimestampNumber.from_timestamp(self.setup_env.start_timestamp, freq)
+        df_generator = self.population_policy.df_generator
+        freq = df_generator.freq
+        start = TimestampNumber.from_timestamp(df_generator.initial_timestamp, freq)
         percent_5 = int(num_rows * percents)
         end_range: TimestampNumber = start + num_rows
         start_range: TimestampNumber = end_range - percent_5
         range = (start_range.to_timestamp(), end_range.to_timestamp())
         return range
-        
-    def time_read_batch(self, storage_info, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs) 
-
-    def time_write_batch(self, storage_info, num_symbols, num_rows):
-        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
-        write_batch_result = self.fresh_lib.write_batch(payloads)
-
-    def time_read_batch_with_columns(self, storage_info, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs_with_cols)
-
-    def peakmem_write_batch(self, storage_info, num_symbols, num_rows):
-        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
-        write_batch_result = self.fresh_lib.write_batch(payloads)
-        # Quick check all is ok (will not affect bemchmarks)
-        assert write_batch_result[0].symbol in self.symbols
-        assert write_batch_result[-1].symbol in self.symbols
-
-    def peakmem_read_batch(self, storage_info, num_symbols, num_rows):
+    
+    def peakmem_read_batch(self, num_symbols, num_rows):
         read_batch_result = self.lib.read_batch(self.read_reqs)
         # Quick check all is ok (will not affect bemchmarks)
         assert read_batch_result[0].data.shape[0] == num_rows
         assert read_batch_result[-1].data.shape[0] == num_rows
 
-    def peakmem_read_batch_with_columns(self, storage_info, num_symbols, num_rows):
+    def time_read_batch(self, num_symbols, num_rows):
+        read_batch_result = self.lib.read_batch(self.read_reqs) 
+
+    def time_write_batch(self, num_symbols, num_rows):
+        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
+        write_batch_result = self.write_lib.write_batch(payloads)
+
+    def time_read_batch_with_columns(self, num_symbols, num_rows):
+        read_batch_result = self.lib.read_batch(self.read_reqs_with_cols)
+
+    def peakmem_write_batch(self, num_symbols, num_rows):
+        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
+        write_batch_result = self.write_lib.write_batch(payloads)
+        # Quick check all is ok (will not affect bemchmarks)
+        assert write_batch_result[0].symbol in self.symbols
+        assert write_batch_result[-1].symbol in self.symbols
+
+    def peakmem_read_batch_with_columns(self, num_symbols, num_rows):
         read_batch_result = self.lib.read_batch(self.read_reqs_with_cols)
         # Quick check all is ok (will not affect bemchmarks)
         assert read_batch_result[0].data.shape[0] == num_rows
         assert read_batch_result[-1].data.shape[0] == num_rows
 
-    def time_read_batch_with_date_ranges(self, storage_info, num_symbols, num_rows):
+    def time_read_batch_with_date_ranges(self, num_symbols, num_rows):
         self.lib.read_batch(self.read_reqs_date_range)
 
-    def peakmem_read_batch_with_date_ranges(self, storage_info, num_symbols, num_rows):
+    def peakmem_read_batch_with_date_ranges(self, num_symbols, num_rows):
         read_batch_result = self.lib.read_batch(self.read_reqs_date_range)
         # Quick check all is ok (will not affect bemchmarks)
         assert read_batch_result[0].data.shape[0] > 2
