+++ python/benchmarks/real_read_write.py
@@ -7,15 +7,11 @@ As of the Change Date specified in that file, in accordance with the Business So
 """
 
 from logging import Logger
-import os
-import time
-from typing import List
 import numpy as np
 import pandas as pd
 
-from arcticdb.util.environment_setup import DataFrameGenerator, LibraryManager, LibraryPopulationPolicy, LibraryType, Storage, get_console_logger, populate_library, populate_library_if_missing
+from arcticdb.util.environment_setup import DataFrameGenerator, LibraryManager, LibraryPopulationPolicy, LibraryType, SequentialDataframesGenerator, Storage, get_console_logger, populate_library, populate_library_if_missing
 from arcticdb.util.utils import DFGenerator, TimestampNumber
-from arcticdb.version_store.library import Library, ReadRequest, WritePayload
 from benchmarks.common import AsvBase
 
 
@@ -158,7 +154,7 @@ class AWSReadWrite(AsvBase):
         self.read_lib.read(symbol=self.symbol, date_range=self.last_20).data
 
 
-class AWSReadWriteWide(AWSReadWrite):
+class AWSWideDataFrameTests(AWSReadWrite):
     """
     This class is for general read write tests on LMDB
 
@@ -185,11 +181,11 @@ class AWSReadWriteWide(AWSReadWrite):
     number_rows= 100
 
     def get_library_manager(self) -> LibraryManager:
-        return AWSReadWriteWide.library_manager
+        return AWSWideDataFrameTests.library_manager
     
     def get_population_policy(self) -> LibraryPopulationPolicy:
-        lpp = LibraryPopulationPolicy(AWSReadWriteWide.params, self.get_logger())
-        lpp.use_parameters_are_columns().set_rows(AWSReadWriteWide.number_rows)
+        lpp = LibraryPopulationPolicy(AWSWideDataFrameTests.params, self.get_logger())
+        lpp.use_parameters_are_columns().set_rows(AWSWideDataFrameTests.number_rows)
         return lpp
     
     def setup_cache(self):
@@ -197,285 +193,3 @@ class AWSReadWriteWide(AWSReadWrite):
         # must implement setup_cache
         super().setup_cache()
 
-
-class AWSListSymbols(AsvBase):
-
-    rounds = 1
-    number = 1 # invoke X times the test runs between each setup-teardown 
-    repeat = 3 # defines the number of times the measurements will invoke setup-teardown
-    min_run_count = 1
-    warmup_time = 0
-
-    timeout = 1200
-    
-    library_manager = LibraryManager(storage=Storage.LMDB, name_benchmark="LIST_SYMBOLS")
-
-    params = [6,8]
-    param_names = ["num_syms"]
-
-    number_columns = 2
-    number_rows = 2
-
-    def get_library_manager(self) -> LibraryManager:
-        return AWSListSymbols.library_manager
-    
-    def get_population_policy(self) -> LibraryPopulationPolicy:
-        lpp = LibraryPopulationPolicy(None, self.get_logger())
-        lpp.set_columns(AWSListSymbols.number_columns)
-        lpp.use_auto_increment_index()
-        return lpp
-
-    def setup_cache(self):
-        manager = self.get_library_manager()
-        assert AWSListSymbols.number == 1, "There must be always one test between setup and tear down"
-        policy = self.get_population_policy()
-        for number_symbols in AWSListSymbols.params:
-            policy.set_parameters([AWSListSymbols.number_rows] * number_symbols)
-            populate_library_if_missing(manager, policy, LibraryType.PERSISTENT, number_symbols)
-
-        manager.log_info() # Always log the ArcticURIs 
-    
-    def setup(self, num_syms):
-        self.population_policy = self.get_population_policy()
-        self.lib = self.get_library_manager().get_library(LibraryType.PERSISTENT, num_syms)
-        self.test_counter = 1
-        assert num_syms == len(self.lib.list_symbols()), "The library contains expected number of symbols"
-        self.lib._nvs.version_store._clear_symbol_list_keys() # clear cache
-
-    def time_list_symbols(self, num_syms):
-        assert self.test_counter == 1, "Test executed only once in setup-teardown cycle" 
-        self.lib.list_symbols()
-        self.test_counter += 1
-
-    def time_has_symbol_nonexisting(self, num_syms):
-        assert self.test_counter == 1, "Test executed only once in setup-teardown cycle" 
-        self.lib.has_symbol("250_sym")        
-        self.test_counter += 1
-
-    def peakmem_list_symbols(self, num_syms):
-        assert self.test_counter == 1, "Test executed only once in setup-teardown cycle" 
-        self.lib.list_symbols()
-        self.test_counter += 1
-
-
-class AWSVersionSymbols(AsvBase):
-
-    rounds = 1
-    number = 3 # invoke X times the test runs between each setup-teardown 
-    repeat = 1 # defines the number of times the measurements will invoke setup-teardown
-    min_run_count = 1
-    warmup_time = 0
-
-    timeout = 1200
-
-    library_manager = LibraryManager(storage=Storage.LMDB, name_benchmark="LIST_SYMBOLS")
-
-    params = [5,7]
-    param_names = ["num_syms"]
-
-    number_columns = 2
-    number_rows = 2
-
-    mean_number_versions_per_symbol = 5
-
-    def get_library_manager(self) -> LibraryManager:
-        return AWSVersionSymbols.library_manager
-    
-    def get_population_policy(self) -> LibraryPopulationPolicy:
-        lpp = LibraryPopulationPolicy(None, self.get_logger())
-        lpp.set_columns(AWSVersionSymbols.number_columns)
-        lpp.use_auto_increment_index()
-        lpp.generate_versions(versions_max=int(1.5 * AWSVersionSymbols.mean_number_versions_per_symbol), 
-                              mean=AWSVersionSymbols.mean_number_versions_per_symbol)
-        lpp.generate_metadata().generate_snapshots()
-        return lpp
-
-    def setup_cache(self):
-        manager = self.get_library_manager()
-        policy = self.get_population_policy()
-        last_snapshot_names_dict = {}
-        for number_symbols in AWSVersionSymbols.params:
-            policy.set_parameters([AWSVersionSymbols.number_rows] * number_symbols)
-            populate_library_if_missing(manager, policy, LibraryType.PERSISTENT, number_symbols)
-            lib = self.get_library_manager().get_library(LibraryType.PERSISTENT, number_symbols)
-            snapshot_name = lib.list_snapshots(load_metadata=False)[-1]
-            last_snapshot_names_dict[number_symbols] = snapshot_name
-        manager.log_info() # Always log the ArcticURIs 
-        return last_snapshot_names_dict
-    
-    def setup(self, last_snapshot_names_dict, num_syms):
-        self.population_policy = self.get_population_policy()
-        self.lib = self.get_library_manager().get_library(LibraryType.PERSISTENT, num_syms)
-        self.test_counter = 1
-        expected_num_versions = AWSVersionSymbols.mean_number_versions_per_symbol * num_syms
-        assert num_syms == len(self.lib.list_symbols()), "The library contains expected number of symbols"
-        assert expected_num_versions - 1 <= len(self.lib.list_versions()), "There are sufficient versions"
-        assert expected_num_versions - 1 <= len(self.lib.list_snapshots()), "There are sufficient snapshots"
-        assert last_snapshot_names_dict[num_syms] is not None
-
-    def time_list_versions(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions()
-
-    def time_list_versions_latest_only(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions(latest_only=True)        
-
-    def time_list_versions_skip_snapshots(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions(skip_snapshots=True)        
-
-    def time_list_versions_latest_only_and_skip_snapshots(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions(latest_only=True, skip_snapshots=True)        
-
-    def time_list_versions_snapshot(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions(snapshot=last_snapshot_names_dict[num_syms])        
-
-    def peakmem_list_versions(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_versions()
-
-    def time_list_snapshots(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_snapshots()
-    
-    def time_list_snapshots_without_metadata(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_snapshots(load_metadata=False)
-
-    def peakmem_list_snapshots(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_snapshots()
-    
-    def peakmem_list_snapshots_without_metadata(self, last_snapshot_names_dict, num_syms):
-        self.lib.list_snapshots(load_metadata=False)
-
-
-class AWSBatchBasicFunctions(AsvBase):
-    """
-    This is similar test to :class:`BatchBasicFunctions`
-    Note that because batch functions are silent we do check if they work correctly along with 
-    peakmem test where this will not influence result in any meaningful way
-    """
-
-    rounds = 1
-    number = 3 # invokes 3 times the test runs between each setup-teardown 
-    repeat = 1 # defines the number of times the measurements will invoke setup-teardown
-    min_run_count = 1
-    warmup_time = 0
-
-    timeout = 1200
-
-    params = [[5, 10], [250, 500]]
-    param_names = ["num_symbols", "num_rows"]
-
-    library_manager = LibraryManager(Storage.LMDB, "BASIC_BATCH")
-
-    number_columns = 10
-
-    def get_library_manager(self) -> LibraryManager:
-        return AWSBatchBasicFunctions.library_manager
-    
-    def get_population_policy(self) -> LibraryPopulationPolicy:
-        lpp = LibraryPopulationPolicy(AWSBatchBasicFunctions.params, self.get_logger())
-        lpp.set_columns(AWSBatchBasicFunctions.number_columns)
-        lpp.use_auto_increment_index()
-        return lpp
-    
-    def setup_cache(self):
-        manager = self.get_library_manager()
-        policy = self.get_population_policy()
-        number_symbols_list, number_rows_list = AWSBatchBasicFunctions.params
-        for number_symbols in number_symbols_list:
-            lib_suffix = number_symbols
-            if not manager.has_library(LibraryType.PERSISTENT, lib_suffix):
-                for number_rows in number_rows_list:
-                    policy.set_parameters([number_rows] * lib_suffix)
-                    # the name of symbols during generation will have now 2 parameters:
-                    # the index of symbol + number of rows
-                    # that allows generating more than one symbol in a library
-                    policy.set_symbol_fixed_str(number_rows) 
-                    populate_library(manager, policy, LibraryType.PERSISTENT, lib_suffix)
-        manager.log_info() # Always log the ArcticURIs 
-
-    def teardown(self, num_symbols, num_rows):
-        # We could clear the modifiable libraries we used
-        self.get_library_manager().clear_all_modifiable_libs_from_this_process()
-
-    def setup(self, num_symbols, num_rows):
-        self.manager = self.get_library_manager()
-        self.population_policy = self.get_population_policy()
-        # We use the same generator as the policy
-
-        self.lib: Library = self.manager.get_library(LibraryType.PERSISTENT, num_symbols)
-        self.write_lib: Library = self.manager.get_library(LibraryType.MODIFIABLE, num_symbols)
-        self.get_logger().info(f"Library {self.lib}") 
-        self.get_logger().debug(f"Symbols {self.lib.list_symbols()}") 
-        
-        # Get generated symbol names
-        self.symbols = []
-        for num_symb_idx in range(num_symbols):
-            # the name is constructed of 2 parts index + number of rows
-            sym_name = self.population_policy.get_symbol_name(num_symb_idx, num_rows)
-            if not self.lib.has_symbol(sym_name):
-                self.get_logger().error(f"symbol not found {sym_name}") 
-            self.symbols.append(sym_name)
-
-        #Construct read requests (will equal to number of symbols)
-        self.read_reqs = [ReadRequest(symbol) for symbol in self.symbols]
-
-        #Construct dataframe that will be used for write requests, not whole DF (will equal to number of symbols)
-        self.df = self.population_policy.df_generator.get_dataframe(num_rows, AWSBatchBasicFunctions.number_columns)
-
-        #Construct read requests based on 2 colmns, not whole DF (will equal to number of symbols)
-        COLS = self.df.columns[2:4]
-        self.read_reqs_with_cols = [ReadRequest(symbol, columns=COLS) for symbol in self.symbols]
-
-        #Construct read request with date_range
-        self.date_range = self.get_last_x_percent_date_range(num_rows, 0.05)
-        self.read_reqs_date_range = [ReadRequest(symbol, date_range=self.date_range) for symbol in self.symbols]
-
-    def get_last_x_percent_date_range(self, num_rows, percents):
-        """
-        Returns a date range tuple selecting last X% of rows of dataframe
-        pass percents as 0.0-1.0
-        """
-        df_generator = self.population_policy.df_generator
-        freq = df_generator.freq
-        start = TimestampNumber.from_timestamp(df_generator.initial_timestamp, freq)
-        percent_5 = int(num_rows * percents)
-        end_range: TimestampNumber = start + num_rows
-        start_range: TimestampNumber = end_range - percent_5
-        range = (start_range.to_timestamp(), end_range.to_timestamp())
-        return range
-    
-    def peakmem_read_batch(self, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs)
-        # Quick check all is ok (will not affect bemchmarks)
-        assert read_batch_result[0].data.shape[0] == num_rows
-        assert read_batch_result[-1].data.shape[0] == num_rows
-
-    def time_read_batch(self, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs) 
-
-    def time_write_batch(self, num_symbols, num_rows):
-        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
-        write_batch_result = self.write_lib.write_batch(payloads)
-
-    def time_read_batch_with_columns(self, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs_with_cols)
-
-    def peakmem_write_batch(self, num_symbols, num_rows):
-        payloads = [WritePayload(symbol, self.df) for symbol in self.symbols]
-        write_batch_result = self.write_lib.write_batch(payloads)
-        # Quick check all is ok (will not affect bemchmarks)
-        assert write_batch_result[0].symbol in self.symbols
-        assert write_batch_result[-1].symbol in self.symbols
-
-    def peakmem_read_batch_with_columns(self, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs_with_cols)
-        # Quick check all is ok (will not affect bemchmarks)
-        assert read_batch_result[0].data.shape[0] == num_rows
-        assert read_batch_result[-1].data.shape[0] == num_rows
-
-    def time_read_batch_with_date_ranges(self, num_symbols, num_rows):
-        self.lib.read_batch(self.read_reqs_date_range)
-
-    def peakmem_read_batch_with_date_ranges(self, num_symbols, num_rows):
-        read_batch_result = self.lib.read_batch(self.read_reqs_date_range)
-        # Quick check all is ok (will not affect bemchmarks)
-        assert read_batch_result[0].data.shape[0] > 2
-        assert read_batch_result[-1].data.shape[0] > 2
