+++ python/benchmarks/realold_finalize_staged_data.py
@@ -8,13 +8,12 @@ As of the Change Date specified in that file, in accordance with the Business So
 
 
 import os
-from arcticdb.util.environment_setup import LibraryManager, LibraryType, Storage
+from arcticdb.util.environment_setup_v1 import Storage, NoSetup
 from arcticdb.util.utils import CachedDFGenerator, TimestampNumber, stage_chunks
 from arcticdb.version_store.library import StagedDataFinalizeMethod
-from benchmarks.common import AsvBase
 
 
-class AWSFinalizeStagedData(AsvBase):
+class AWSFinalizeStagedData:
     """
     Checks finalizing staged data. Note, that staged symbols can be finalized only twice,
     therefore certain design decisions must be taken in advance so each process sets up
@@ -29,16 +28,11 @@ class AWSFinalizeStagedData(AsvBase):
 
     timeout = 1200
 
+    SETUP_CLASS = (NoSetup(storage=Storage.AMAZON, 
+                                                      prefix="FINALIZE"))
+                   
     params = [500, 1000] # Test data [10, 20]
     param_names = ["num_chunks"]
-
-    library_manager = LibraryManager(Storage.LMDB, "FINALIZE")
-
-    def get_library_manager(self) -> LibraryManager:
-        return AWSFinalizeStagedData.library_manager
-    
-    def get_population_policy(self):
-        pass
 
     def setup_cache(self):
         # Preconditions for this test
@@ -47,18 +41,22 @@ class AWSFinalizeStagedData(AsvBase):
         assert AWSFinalizeStagedData.rounds == 1
         assert AWSFinalizeStagedData.warmup_time == 0
 
-        manager = self.get_library_manager()
-        manager.clear_all_modifiable_libs()
-        manager.log_info()
-
+        set_env = AWSFinalizeStagedData.SETUP_CLASS
+        info = set_env.get_storage_info()
+        set_env.logger().info(f"storage info object: {info}")
         df_cache = CachedDFGenerator(500000, [5])
-        return df_cache
+        cache = {
+            "storage_info" : info,
+            "df_cache" : df_cache
+        }
+        return cache
     
     def setup(self, cache, num_chunks: int):
-        self.df_cache: CachedDFGenerator = cache
-        self.logger = self.get_logger()
+        self.df_cache: CachedDFGenerator = cache["df_cache"]
+        self.set_env = NoSetup.from_storage_info(cache["storage_info"])
 
-        self.lib = self.get_library_manager().get_library(LibraryType.MODIFIABLE)
+        self.pid = os.getpid()
+        self.lib = self.set_env.get_modifiable_library(self.pid)
 
         INITIAL_TIMESTAMP: TimestampNumber = TimestampNumber(
             0, self.df_cache.TIME_UNIT
@@ -66,22 +64,22 @@ class AWSFinalizeStagedData(AsvBase):
 
         df = self.df_cache.generate_dataframe_timestamp_indexed(200, 0, self.df_cache.TIME_UNIT)
         list_of_chunks = [10000] * num_chunks
-        self.symbol = f"symbol_{os.getpid()}"
+        self.symbol = self.set_env.get_symbol_name_template(self.pid)
 
         self.lib.write(self.symbol, data=df, prune_previous_versions=True)
         stage_chunks(self.lib, self.symbol, self.df_cache, INITIAL_TIMESTAMP, list_of_chunks)
 
     def teardown(self, cache: CachedDFGenerator, param: int):
-        self.get_library_manager().clear_all_modifiable_libs_from_this_process()
+        self.set_env.delete_modifiable_library(self.pid)    
 
     def time_finalize_staged_data(self, cache: CachedDFGenerator, param: int):
-        self.logger.info(f"Library: {self.lib}")
-        self.logger.info(f"Symbol: {self.symbol}")
+        self.set_env.logger().info(f"Library: {self.lib}")
+        self.set_env.logger().info(f"Symbol: {self.symbol}")
         assert self.symbol in self.lib.get_staged_symbols()
         self.lib.finalize_staged_data(self.symbol, mode=StagedDataFinalizeMethod.WRITE)
 
     def peakmem_finalize_staged_data(self, cache: CachedDFGenerator, param: int):
-        self.logger.info(f"Library: {self.lib}")
-        self.logger.info(f"Symbol: {self.symbol}")
+        self.set_env.logger().info(f"Library: {self.lib}")
+        self.set_env.logger().info(f"Symbol: {self.symbol}")
         assert self.symbol in self.lib.get_staged_symbols()
         self.lib.finalize_staged_data(self.symbol, mode=StagedDataFinalizeMethod.WRITE)
